Basic concept for authorization (#824)
* Basic concept for multi-tenancy

* Added authz for ListCloudServices

* Adding access types

* starting with authorization of assessment results

* migrating assesment result functions to seperate go files

* Added authz to metric configuration requests

+++ service/orchestrator/cloud_service.go
@@ -75,9 +75,11 @@ func (s *Service) RegisterCloudService(_ context.Context, req *orchestrator.Regi
 }
 
 // ListCloudServices implements method for OrchestratorServer interface for listing all cloud services
-func (svc *Service) ListCloudServices(_ context.Context, req *orchestrator.ListCloudServicesRequest) (
+func (svc *Service) ListCloudServices(ctx context.Context, req *orchestrator.ListCloudServicesRequest) (
 	res *orchestrator.ListCloudServicesResponse, err error) {
-	res = new(orchestrator.ListCloudServicesResponse)
+	var conds []any
+	var allowed []string
+	var all bool
 
 	// Validate tne request
 	if err = api.ValidateListRequest[*orchestrator.CloudService](req); err != nil {
@@ -87,9 +89,18 @@ func (svc *Service) ListCloudServices(_ context.Context, req *orchestrator.ListC
 		return
 	}
 
+	res = new(orchestrator.ListCloudServicesResponse)
+
+	// Retrieve list of allowed cloud service according to our authorization strategy. No need to specify any conditions
+	// to our storage request, if we are allowed to see all cloud services.
+	all, allowed = svc.authz.AllowedCloudServices(ctx)
+	if !all {
+		conds = append(conds, allowed)
+	}
+
 	// Paginate the cloud services according to the request
 	res.Services, res.NextPageToken, err = service.PaginateStorage[*orchestrator.CloudService](req, svc.storage,
-		service.DefaultPaginationOpts)
+		service.DefaultPaginationOpts, conds...)
 	if err != nil {
 		return nil, status.Errorf(codes.Internal, "could not paginate results: %v", err)
 	}
@@ -98,12 +109,17 @@ func (svc *Service) ListCloudServices(_ context.Context, req *orchestrator.ListC
 }
 
 // GetCloudService implements method for OrchestratorServer interface for getting a cloud service with provided id
-func (s *Service) GetCloudService(_ context.Context, req *orchestrator.GetCloudServiceRequest) (response *orchestrator.CloudService, err error) {
+func (s *Service) GetCloudService(ctx context.Context, req *orchestrator.GetCloudServiceRequest) (response *orchestrator.CloudService, err error) {
 	if req == nil {
 		return nil, status.Errorf(codes.InvalidArgument, api.ErrRequestIsNil.Error())
 	}
 	if req.CloudServiceId == "" {
 		return nil, status.Errorf(codes.InvalidArgument, orchestrator.ErrIDIsMissing.Error())
+	}
+
+	// Check, if this request has access to the cloud service according to our authorization strategy.
+	if !s.authz.CheckAccess(ctx, service.AccessRead, req) {
+		return nil, service.ErrPermissionDenied
 	}
 
 	response = new(orchestrator.CloudService)
@@ -125,6 +141,11 @@ func (s *Service) UpdateCloudService(ctx context.Context, req *orchestrator.Upda
 
 	if req.CloudServiceId == "" {
 		return nil, status.Errorf(codes.InvalidArgument, "service id is empty")
+	}
+
+	// Check, if this request has access to the cloud service according to our authorization strategy.
+	if !s.authz.CheckAccess(ctx, service.AccessUpdate, req) {
+		return nil, service.ErrPermissionDenied
 	}
 
 	count, err := s.storage.Count(req.Service, "id = ?", req.CloudServiceId)
@@ -150,9 +171,14 @@ func (s *Service) UpdateCloudService(ctx context.Context, req *orchestrator.Upda
 }
 
 // RemoveCloudService implements method for OrchestratorServer interface for removing a cloud service
-func (s *Service) RemoveCloudService(_ context.Context, req *orchestrator.RemoveCloudServiceRequest) (response *emptypb.Empty, err error) {
+func (s *Service) RemoveCloudService(ctx context.Context, req *orchestrator.RemoveCloudServiceRequest) (response *emptypb.Empty, err error) {
 	if req.CloudServiceId == "" {
 		return nil, status.Errorf(codes.InvalidArgument, "service id is empty")
+	}
+
+	// Check, if this request has access to the cloud service according to our authorization strategy.
+	if !s.authz.CheckAccess(ctx, service.AccessDelete, req) {
+		return nil, service.ErrPermissionDenied
 	}
 
 	err = s.storage.Delete(&orchestrator.CloudService{Id: req.CloudServiceId})
